refactor add room form
